requirements.txt
google-generativeai==0.8.3
openai==1.52.2
python-dotenv==1.0.1
